Add entity states for events
What are entities? Entities are any kind of data class(as defined by the discord.py API docs) that has an ID associated with it, thus making them unique from events. Events don't actually record actions.

Think of this as a low level audit log for sigma. This will allow a lot more types of plugins to be created. One example is checking that last kicker of Member. When a state is created, a database collection is also created automatically as needed.